refactor: Split payment_orchestrator.go into logical modules
Split the 1403-line payment_orchestrator.go into four focused files:

- payment_orchestrator.go (202 lines): struct, config, constructor, errors,
  event publishing
- payment_orchestrator_saga.go (464 lines): saga orchestration, state
  transitions, failure handling, bucket evaluation
- payment_orchestrator_ledger.go (532 lines): double-entry ledger posting
  for both standard and internal clearing flows
- payment_orchestrator_lien.go (279 lines): async lien execution with
  exponential backoff retry and distributed locking

All files remain in package service with methods on *PaymentOrchestrator,
preserving full API compatibility. No behavioral changes - all existing
tests pass without modification.
